Add files via upload
A small R analysis was run on a movie dataset, We calculated distinct no of Users & movies. We cleaned the data and split it into a transactional database plotted item frequency plot for the top 10 movies. 
Discovered the most popular & least popular items using the apriori function in support of 0.3 & min size of 2. 
Summarised the set of extracted rules with a support of 30% & min confidence of 90%.